perf: update open red-envelope status
